refactor: extract fused-hybrid trunk runtime (P2)
Moves the fused-hybrid trunk implementation out of the engine monolith
into Sources/RealModelInference/FusedHybridRuntime.swift via a frozen
engine extension:

- ensureFusedHybridCompiled + compileFusedHybridLayers
- generateIncrementalFusedHybridLlama loop
- per-trunk state stays as flat internal engine fields (LayerStorage is
  ~Copyable so bundling into one session struct is not expressible)

Engine internals widened private->internal where the trunk file reads
them (config/policies/tokenizer/assets, sampling + embedding helpers).
No behavior change.

Verified: golden replay bit-identical incl. qwen15b fused trace; full
suite green.
